Added a special launcher activity
Right now, this just allows me to easily make "CPAN" the name of the
application, but immediately load the "CPAN Module Search" activity, which is a
terrible application name.

In the future, this will probably be used to determine which activity to launch
at startup based upon type of device, previous state, user preference, etc.