Working towards better HTTP client management
To avoid setting up and tearing down the HTTP client object, we try to share
them. However, the previous solution was very manual and difficult to keep up
with. This adds a reference counter to keep better track of pending actions.
However, it still is not working quite right in every case.