Still working on making the http client stuff robust
Added some tight timeout parameters to Gravatar requests because we really don't
care about those. Made it so that the indeterminate progress shows anytime the
http client manager is busy in the search.

Plan to do the same for the viewer as well.

I suspect that I'm running into some sort of deadlock updating the Gravatar
images now, though. Too many synchronized actions...