Add a reaper to reap long running Gravatar requests
The Gravatar server takes the connections and keeps them alive, but returns
nothing. I would rather just drop the connection after 10 seconds than have the
Gravatar images, so this is what we now do.

Ultimately, we should keep an LRU cache of Gravatar images so we can pick up any
missed ones on a later retry.