First version of the app with a working tablet view
This adds a working version of the search activity that displayes the search
results on the left and the select item on the right. It does not look right as
the header for the view blends with the ActionBar and the search should be moved
into the action bar.

To be done:

 * Separate the header into a floating box, distinct from the ActionBar
 * Show the search in the action bar instead of the search sidebar
 * Keep items selected in the search results when displaying in the view
 * Add a triangle/arrow drawable to show which item is selected, similar to how
   the Android design guidelines demonstrate
 * Show a blank box telling the user to search for something when there are no
   results.
 * Automatically select the first result when search returns something