Tablet-sized display was causing AsyncTask rejections
The problem was that each Gravatar icon was being fetched in a separate thread.
Since no test of mine had caused more than 10 tasks to be loaded at once, this
was not a problem. However, the tablet sized screen made more than 10 rows
loading at once common, so this exception was starting to happen with some
frequency. Fixed by loading all the Gravatar icons for each search in one go.