Improving the search in Eclair, breaking Honeycomb
This moves the search out of the module search fragment and makes it part of the
activity itself. This also makes it so that the module search activity is better
integrated with Android's built-in search features, though I am not using the
built-in search dialog at this point because I want the search EditText to stay
up all the time.

This change adds the search menu to the ActionBar for Honeycomb, but does not
implement it yet, so search is now broken on Honeycomb.