Various fragment-related fixes
Nesting the module info fragment seemed like a good idea, but it caused massive trauma, so I made it a layout include instead.

I also had trouble with swapping the help/placeholder fragment with the POD view. Apparently, Android really hates it if you code that in as a literal XML resource and then change it programmatically. It is much happier if you just stick to either keeping it permanent via XML or setting it up and swapping it programmatically. At least, that seems to be easier. If there's some magic incantation to make it work, I couldn't find it and my friends over on StackOverflow recommended against it.