Upgrades to Eclipse/Android SDK and embedding the support library
Part of this work was trying to get a test project for the application, but
after several hours of trying to get the test project setup, I've given up for
the time being. For some reason, the test project cannot find the support
library code even though I do everything the Dev Guide suggests and I've tried
the suggestions of everyone I can find on Stackoverflow and elsewhere without
success. The usual "setup the support lib as an export in the main project and
add that as a dependency definitely does not help me."