Improve the event handling code, focused on fetch rather than changes to resultset
In some ways I would prefer to handle the updates on any change to the result
set or the instances it refers to. However, in reality, this application will
never update these unless there's a fetch, so tying the events to the fetching
process is simpler and a better description of how this app works.