Issue #11: Fixes Gravatar took too long RuntimeException problem
This limits the scope of when this exception can escape it's try block by
canceling the timer immediately after we get a response back from the Gravatar
server.

Then, I added another try block to guarantee that the GRAVATAR_TOOK_TOO_LONG
exception is caught and squelched regardless of what weird order it might happen
in.